Re-organize HARP runtime.
