--- sources/lib/run-time/linux-exceptions.c
@@ -1,3 +1,0 @@
-#ifdef X86_LINUX_PLATFORM
-#include "x86-linux-exceptions.c"
-#endif
