--- sources/lib/run-time/pentium-freebsd/x86-linux-threads-primitives.c
@@ -1,9 +1,0 @@
-#define _GNU_SOURCE
-
-/* flag to select Linux */
-#define LINUX_PLATFORM
-#define X86_LINUX_PLATFORM
-//#define BOEHM_GC
-
-
-#include "linux-threads-primitives.c"
