+++ sources/lib/run-time/unix-collector.c
@@ -5,11 +5,6 @@
 
 #define _GNU_SOURCE
 
-/* flag to select Linux */
-#define LINUX_PLATFORM
-#define X86_LINUX_PLATFORM
-//#define BOEHM_GC
-
 #include <sys/signal.h>
 
 #include "collector.c"
